# internal/docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /v2/legends/daily-tracking:
    get:
      description: Returns per-day legend stats for all players in a guild over a
        date range.
      parameters:
      - description: Discord guild ID
        in: query
        name: guild_id
        required: true
        type: integer
      - description: Start date in YYYY-MM-DD format
        in: query
        name: start_date
        required: true
        type: string
      - description: End date in YYYY-MM-DD format
        in: query
        name: end_date
        required: true
        type: string
      - description: Restrict to a single clan
        in: query
        name: clan_tag
        type: string
      - description: Maximum number of players to return
        in: query
        name: limit
        type: integer
      - description: Number of players to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get legends daily tracking
      tags:
      - Bot Legends Endpoints
  /v2/legends/guild-stats:
    get:
      description: Returns legends leaderboard and player stats for a guild.
      parameters:
      - description: Discord guild ID
        in: query
        name: guild_id
        required: true
        type: integer
      - description: Season (YYYY-MM)
        in: query
        name: season
        type: string
      - description: Max number of top players to return
        in: query
        name: limit_top_players
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get guild legends statistics
      tags:
      - Bot Legends Endpoints
...
swagger: "2.0"